requirements.txt
fastapi
uvicorn
python-dotenv
gtts
python-multipart
google.generativeai